:bug: domain값이 가져와지지 않고 공백인 버그 수정
