+++ src/main/kotlin/team/themoment/gsmNetworking/domain/auth/controller/AuthController.kt
@@ -4,7 +4,8 @@ import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import team.themoment.gsmNetworking.common.cookie.CookieUtil
+import team.themoment.gsmNetworking.common.cookie.CookieManager
+import team.themoment.gsmNetworking.common.cookie.impl.CookieUtil
 import team.themoment.gsmNetworking.domain.auth.service.ReissueTokenService
 import team.themoment.gsmNetworking.global.security.jwt.properties.JwtProperties
 import javax.servlet.http.HttpServletRequest
@@ -13,14 +14,15 @@ import javax.servlet.http.HttpServletResponse
 @RestController
 @RequestMapping("api/v1/auth")
 class AuthController(
-    private val reissueTokenService: ReissueTokenService
+    private val reissueTokenService: ReissueTokenService,
+    private val cookieManager: CookieManager
 ) {
 
     @PatchMapping("/reissue")
     fun reissueToken(request: HttpServletRequest, response: HttpServletResponse): ResponseEntity<Unit> {
-        val refreshToken = CookieUtil.getCookieValueOrNull(request.cookies, JwtProperties.REFRESH)
+        val refreshToken = cookieManager.getCookieValueOrNull(request.cookies, JwtProperties.REFRESH)
         val tokenDto = reissueTokenService.execute(refreshToken)
-        CookieUtil.addTokenCookie(tokenDto, response)
+        cookieManager.addTokenCookie(tokenDto, response)
         return ResponseEntity.noContent().build()
     }
