+++ src/main/kotlin/team/themoment/gsmNetworking/global/filter/TokenRequestFilter.kt
@@ -3,7 +3,7 @@ package team.themoment.gsmNetworking.global.filter
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Component
 import org.springframework.web.filter.OncePerRequestFilter
-import team.themoment.gsmNetworking.common.cookie.CookieUtil
+import team.themoment.gsmNetworking.common.cookie.CookieManager
 import team.themoment.gsmNetworking.global.security.jwt.TokenParser
 import team.themoment.gsmNetworking.global.security.jwt.properties.JwtProperties
 import javax.servlet.FilterChain
@@ -12,7 +12,8 @@ import javax.servlet.http.HttpServletResponse
 
 @Component
 class TokenRequestFilter(
-    private val tokenParser: TokenParser
+    private val tokenParser: TokenParser,
+    private val cookieManager: CookieManager
 ) : OncePerRequestFilter() {
 
     override fun doFilterInternal(
@@ -21,7 +22,7 @@ class TokenRequestFilter(
         filterChain: FilterChain
     ) {
         if (request.cookies != null){
-            val accessToken = CookieUtil.getCookieValueOrNull(cookies = request.cookies, name = JwtProperties.ACCESS)
+            val accessToken = cookieManager.getCookieValueOrNull(cookies = request.cookies, name = JwtProperties.ACCESS)
 
             if (!accessToken.isNullOrEmpty()) {
                 val authentication = tokenParser.authentication(accessToken)
